+++ api/nameguard/checks/grapheme/invisible.py
@@ -1,12 +1,12 @@
 from label_inspector.models import InspectorGraphemeWithConfusablesResult as Grapheme
-from nameguard.models import CheckStatus, Check, GenericCheckResult, GraphemeCheckResult
-
+from nameguard.models import CheckStatus, Check, GenericCheckResult, GraphemeCheckResult, UNINSPECTED_SKIP_MESSAGE
 
 STATUS = CheckStatus.ALERT
 
 # title: No Hidden Characters
 TITLE_PASS = 'Character Visibility'
 TITLE_FAIL = 'Character Visibility'
+TITLE_SKIP = 'Character Visibility'
 
 G_MESSAGE_PASS = 'Visible'
 L_MESSAGE_PASS = 'All characters are visible'
@@ -27,3 +27,13 @@ def check_grapheme(grapheme: Grapheme) -> GenericCheckResult:
         _name_message=N_MESSAGE_PASS if passed else N_MESSAGE_FAIL,
         _title=TITLE_PASS if passed else TITLE_FAIL,
     )
+
+
+UNINSPECTED_SKIP_CHECK_RESULT = GraphemeCheckResult(
+    check=Check.INVISIBLE,
+    status=CheckStatus.SKIP,
+    _grapheme_message=UNINSPECTED_SKIP_MESSAGE,
+    _label_message=UNINSPECTED_SKIP_MESSAGE,
+    _name_message=UNINSPECTED_SKIP_MESSAGE,
+    _title=TITLE_SKIP,
+)