+++ api/tests/test_api.py
@@ -8,7 +8,7 @@ import os
 from dotenv import load_dotenv
 
 from nameguard.models import FakeEthNameCheckStatus
-from nameguard.utils import labelhash_from_label
+from nameguard.utils import labelhash_from_label, MAX_INSPECTED_NAME_CHARACTERS, MAX_NUMBER_OF_NAMES_IN_BULK
 
 load_dotenv()  # run lambda api tests if LAMBDA_ROOT_URL is set
 running_lambda_tests = bool(os.environ.get('LAMBDA_ROOT_URL'))
@@ -225,19 +225,22 @@ def test_bulk_inspect_name_post(test_client):
                 'beautiful_name',
                 'subtitle',
                 'title',
+                'inspected',
             }
             for r in res_json['results']
         ]
     )
 
-    # more than 250 names
-    response = test_client.post('/bulk-inspect-names', json={'names': names * 126})
+    # more than MAX_NUMBER_OF_NAMES_IN_BULK names
+    names_to_inspect = names * 126
+    assert len(names_to_inspect) > MAX_NUMBER_OF_NAMES_IN_BULK
+    response = test_client.post('/bulk-inspect-names', json={'names': names_to_inspect})
     assert response.status_code == 422
 
 
 @pytest.mark.flaky(retries=2, condition=not pytest.use_monkeypatch)
 def test_bulk_inspect_name_post_stress(test_client):
-    names = [f'[{labelhash_from_label(str(i))[2:]}].eth' for i in range(250)]
+    names = [f'[{labelhash_from_label(str(i))[2:]}].eth' for i in range(MAX_NUMBER_OF_NAMES_IN_BULK)]
     print(names)
     response = test_client.post('/bulk-inspect-names', json={'names': names, 'network_name': 'mainnet'})
     assert response.status_code == 200
@@ -861,6 +864,19 @@ def test_primary_name_get_unnormalized(test_client):
     assert res_json['display_name'] == 'Unnamed fa9a'
 
 
+@pytest.mark.flaky(retries=2, condition=not pytest.use_monkeypatch)
+def test_primary_name_get_uninspected(test_client):
+    address = '0xf4A4D9C75dA65d507cfcd5ff0aCB73D40D3A3bCB'
+    response = test_client.get(f'/secure-primary-name/mainnet/{address}')
+    assert response.status_code == 200
+    res_json = response.json()
+    print(res_json)
+    assert res_json['impersonation_status'] is None
+    assert res_json['primary_name_status'] == 'uninspected'
+    assert res_json['primary_name'] is None
+    assert res_json['display_name'] == 'Unnamed f4a4'
+
+
 def test_primary_name_get_invalid_address(test_client):
     address = '0xfA9A134f997b3d48e122d043E12d04E909b1107g'
     response = test_client.get(f'/secure-primary-name/mainnet/{address}')
@@ -1256,3 +1272,57 @@ def test_fake_eth_name_check_fields_missing_title(test_client, contract_address,
     }
     response = test_client.post('/fake-eth-name-check', json=json_req)
     assert response.status_code == 422
+
+
+def test_inspect_name_post_long(test_client):
+    name = '≡ƒÿ║' * 50  # equal to MAX_INSPECTED_NAME_CHARACTERS
+    assert len(name) == MAX_INSPECTED_NAME_CHARACTERS
+
+    response = test_client.post('/inspect-name', json={'name': name, 'network_name': 'mainnet'})
+    assert response.status_code == 200
+    res_json = response.json()
+    assert res_json['highest_risk']['check'] == 'normalized'
+    assert res_json['inspected']
+    pprint(res_json)
+
+
+def test_bulk_inspect_name_post_long(test_client):
+    name = '≡ƒÿ║' * 50  # equal to MAX_INSPECTED_NAME_CHARACTERS
+    assert len(name) == MAX_INSPECTED_NAME_CHARACTERS
+
+    names = [name] * MAX_NUMBER_OF_NAMES_IN_BULK
+    response = test_client.post('/bulk-inspect-names', json={'names': names, 'network_name': 'mainnet'})
+    assert response.status_code == 200
+    res_json = response.json()
+    for x in res_json['results']:
+        assert x['highest_risk']['check'] == 'normalized'
+        assert x['inspected']
+
+
+def test_bulk_inspect_name_post_too_long(test_client):
+    name = '≡ƒÿ║' * 51  # more than MAX_INSPECTED_NAME_CHARACTERS
+    assert len(name) > MAX_INSPECTED_NAME_CHARACTERS
+
+    names = [name] * MAX_NUMBER_OF_NAMES_IN_BULK
+    response = test_client.post('/bulk-inspect-names', json={'names': names, 'network_name': 'mainnet'})
+    assert response.status_code == 200
+    res_json = response.json()
+
+    for x in res_json['results']:
+        assert x['highest_risk']['check'] == 'uninspected'
+        assert x['normalization'] == 'unnormalized'
+        assert not x['inspected']
+
+
+def test_bulk_inspect_name_post_too_long_normalized(test_client):
+    name = 'abcd' * 51  # more than MAX_INSPECTED_NAME_CHARACTERS
+    assert len(name) > MAX_INSPECTED_NAME_CHARACTERS
+
+    names = [name] * MAX_NUMBER_OF_NAMES_IN_BULK
+    response = test_client.post('/bulk-inspect-names', json={'names': names, 'network_name': 'mainnet'})
+    assert response.status_code == 200
+    res_json = response.json()
+
+    for x in res_json['results']:
+        assert x['highest_risk']['check'] == 'uninspected'
+        assert x['normalization'] == 'normalized'