+++ apps/nameguard.io/src/app/docs/report/page.tsx
@@ -398,6 +398,7 @@ const getExampleReportData = (rating: Rating): ConsolidatedNameGuardReport => {
         title: "Looks Good",
         subtitle: "All security checks passed!",
         beautiful_name: name.displayName,
+        inspected: true,
       };
     case Rating.warn:
       return {
@@ -415,6 +416,7 @@ const getExampleReportData = (rating: Rating): ConsolidatedNameGuardReport => {
         title: "Some Risk",
         subtitle: "Review risks before proceeding",
         beautiful_name: name.displayName,
+        inspected: true,
       };
     case Rating.alert:
       return {
@@ -432,6 +434,7 @@ const getExampleReportData = (rating: Rating): ConsolidatedNameGuardReport => {
         title: "High Risk",
         subtitle: "Better not to use this name",
         beautiful_name: name.displayName,
+        inspected: true,
       };
   }
 };